🔍 refactor.error element 정리
mui icon을 스타일로 빼놓긴했는데 적용되는지 몰라서 원래 인라인 코드를 건들지는 않았습니다
